beginner_tutorials/src/requirements.txt
absl-py==0.10.0
actionlib==1.11.13
angles==1.9.12
apturl==0.5.2
astor==0.8.1
backports.weakref==1.0rc1
base-local-planner==1.14.7
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
bondpy==1.8.3
Brlapi==0.6.4
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
capabilities==0.2.0
catkin==0.7.20
catkin-pkg==0.4.20
chardet==2.3.0
checkbox-support==0.22
command-not-found==0.3
controller-manager==0.13.5
controller-manager-msgs==0.13.5
create-dashboard==2.3.1
create-driver==2.3.1
create-node==2.3.1
cryptography==1.2.3
cv-bridge==1.12.8
cycler==0.10.0
defer==1.0.6
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
distro==1.5.0
docopt==0.6.2
docutils==0.16
dynamic-reconfigure==1.5.50
feedparser==5.1.3
gast==0.4.0
gazebo-plugins==2.5.20
gazebo-ros==2.5.20
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.14
gps-common==0.3.0
grpcio==1.31.0
guacamole==0.9.2
h5py==2.10.0
html5lib==0.9999999
httplib2==0.9.1
idna==2.0
image-geometry==1.12.8
importlib-metadata==1.7.0
interactive-markers==1.11.5
Jinja2==2.8
joint-state-publisher==1.12.15
joint-state-publisher-gui==1.12.15
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
kobuki-dashboard==0.5.7
language-selector==0.1
laser-geometry==1.6.5
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==3.2.2
MarkupSafe==0.23
matplotlib==3.0.3
message-filters==1.12.15
moveit-commander==0.9.18
moveit-ros-planning-interface==0.9.18
moveit-ros-visualization==0.9.18
numpy==1.18.5
oauthlib==1.0.3
onboard==1.2.0
opencv-contrib-python==4.4.0.42
opencv-contrib-python-headless==4.4.0.42
opencv-python-headless==4.4.0.42
openni2-launch==0.4.2
padme==1.1.1
pexpect==4.0.1
Pillow==3.1.2
pipreqs==0.4.10
plainbox==0.25
pluginlib==1.11.3
protobuf==3.13.0
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.4.7
pyserial==3.4
python-apt==1.1.0b1+ubuntu0.16.4.9
python-dateutil==2.8.1
python-debian==0.1.27
python-qt-binding==0.3.7
python-systemd==231
pyxdg==0.25
PyYAML==5.3.1
qt-dotgraph==0.3.17
qt-gui==0.3.17
qt-gui-cpp==0.3.17
qt-gui-py-common==0.3.17
reportlab==3.3.0
requests==2.9.1
resource-retriever==1.12.6
rocon-app-manager==0.9.1
rocon-app-utilities==0.9.1
rocon-apps==0.9.1
rocon-console==0.3.2
rocon-ebnf==0.3.2
rocon-gateway==0.8.1
rocon-gateway-utils==0.8.1
rocon-hub==0.8.1
rocon-hub-client==0.8.1
rocon-interactions==0.3.2
rocon-master-info==0.3.2
rocon-python-comms==0.3.2
rocon-python-redis==0.3.2
rocon-python-utils==0.3.2
rocon-python-wifi==0.3.2
rocon-semantic-version==0.3.2
rocon-uri==0.3.2
rosapi==0.11.9
rosbag==1.12.15
rosboost-cfg==1.14.6
rosbridge-library==0.11.9
rosbridge-server==0.11.9
rosclean==1.14.6
roscreate==1.14.6
rosgraph==1.12.15
roslaunch==1.12.15
roslib==1.14.6
roslint==0.11.0
roslz4==1.12.15
rosmake==1.14.6
rosmaster==1.12.15
rosmsg==1.12.15
rosnode==1.12.15
rosparam==1.12.15
rospkg==1.2.6
rospy==1.12.15
rospy-message-converter==0.5.2
rosserial-python==0.7.7
rosservice==1.12.15
rostest==1.12.15
rostopic==1.12.15
rosunit==1.14.6
roswtf==1.12.15
rqt-action==0.4.9
rqt-bag==0.4.14
rqt-bag-plugins==0.4.14
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.12.17
sensor-msgs==1.12.7
sessioninstaller==0.0.0
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smbus==1.1.post2
smclib==1.8.3
srdfdom==0.4.2
swri-transform-util==2.14.0
system-service==0.3
tensorboard==1.12.2
tensorflow==1.2.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
tf2-sensor-msgs==0.5.20
topic-tools==1.12.15
turtlebot3-automatic-parking==1.1.0
turtlebot3-automatic-parking-vision==1.1.0
turtlebot3-autorace-control==1.2.0
turtlebot3-autorace-core==1.2.0
turtlebot3-example==1.2.1
turtlebot3-teleop==1.2.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unique-id==1.0.5
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urdfdom-py==0.3.3
urllib3==1.13.1
usb-creator==0.3.0
Werkzeug==1.0.1
world-canvas-server==0.2.0
world-canvas-utils==0.2.0
xacro==1.11.3
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
yarg==0.1.9
zipp==1.2.0
